+++ datadog/CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.0.1 - beta
 ## Added
 - Fixed issue with Datadog Tags configuration parameter.
+- Fixed hostname issue in push functionality.
 
 # 1.0.0
 ## Added